refactor: extract regex test logic from RuleEditSheet to onTest callback, fix unlocalized PatternSyntaxException
